Added autofac extension, refactoring

+++ Lykke.B2c2Client/Models/Rest/OrderRequest.cs
@@ -17,10 +17,10 @@ namespace Lykke.B2c2Client.Models.Rest
         public Side Side { get; set; }
 
         [JsonProperty("price")]
-        public double Price { get; set; }
+        public decimal Price { get; set; }
 
         [JsonProperty("quantity")]
-        public double Quantity { get; set; }
+        public decimal Quantity { get; set; }
 
         [JsonProperty("order_type"), JsonConverter(typeof(StringEnumConverter))]
         public OrderType OrderType { get; set; }
